SPARKC-52: Ignore SaveMode bubbles up Rx error.
Motivation
----------
The Ignore save modes work as expected but in turn they return an
empty observable. The code uses .last() to wait for the batch to
complete but if only ever an empty() observable is returned a
NoSuchElementException is raised.

Modifications
-------------
Since the underlying method returns a Unit (void) anyways, it is
fine to use lastOrElse(null) to suppress any errors in this regard.

Change-Id: I38040674daccef231fd32a491c9453db72cefc8f
Reviewed-on: http://review.couchbase.org/64813
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Michael Nitschinger <michael@nitschinger.at>