SPARKC-71: Properly handle custom nested schema.
A user reported issues when providing a custom schema
which is nested and the fields are not ordered alphabetically.
Together with the N1QL output this only did work on the first
level but as soon as nesting comes into the picture the
code fails at runtime.

The fix relies on the spark primitives for json decoding
now completely instead of partially rolling our own logic
making it more flexible and actually less complicated in
our N1QLRelation.

Change-Id: I98eb60fc853aaeb612ee4b3bcb803ff69b6900a9
Reviewed-on: http://review.couchbase.org/70463
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Michael Nitschinger <michael@nitschinger.at>