SPARKC-71: Pre-filter loaded dataframe.
Motivation
----------
For some reason, if the dataframe is not filtered to the selected
columns before turning it into an RDD and sending it back to Spark
the RDD does not come out at the user level with all the schema
columns, which leads to inconsistent behavior when selecting
only a couple fields.

This might be a bug in spark, but I'm not sure at this point (or
if its just the way its supposed to work).

Modifications
-------------
Preprocess the dataframe with the selected field columsn as long as
the right schema is in scope, so that when passed up the stack everything
is already in the right place.

Result
------
Projections work.

Change-Id: Ida7394c0394a459410cf4d643582540422f9fdbd
Reviewed-on: http://review.couchbase.org/72273
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Michael Nitschinger <michael@nitschinger.at>