SPARKC-73: Allow to also store the ID field in the doc.
Motivation
----------
It has been raised that it should be possible to not remove the
document ID from the body on DF write since it might be needed
in later processing.

Modifications
-------------
This changeset adds another property "removeIdField" which can be
set to false (defaults to true). If false, the document ID field
will not be removed from the content body.

Change-Id: I5656a92eaacff5dd4480835056a3e49591609845
Reviewed-on: http://review.couchbase.org/73569
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Michael Nitschinger <michael@nitschinger.at>