SPARKC-80: Add RBAC support
Motivation
----------
This changeset adds support for Server 5.0 and RBAC by allowing
to set the username and password on the config instead of just
on a per bucket level.

Modifications
-------------
In order to make this work the dcp-client and the other dependencies
need to be bumped to support the new feature, leading to couple of
related cascading changes.

Result
------
5.0 RBAC support

Change-Id: I313bccaab7422f0617d4fbf91b49396a8d4549b1
Reviewed-on: http://review.couchbase.org/80050
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Michael Nitschinger <michael@nitschinger.at>