SPARKC-77: allow to easily configure global timeouts.
Motivation
----------
In addition to allowing per-operation timeout overrides in a separate
commit, this changeset allows to set the default/global timeout
settings from the spark config instead of having to resort to
the system property approach.

Modifications
-------------
The CouchbaseConfig accepts the property names (identical to the
CouchbaseEnvironment) and parses/forwards them into the connection
where if set are applied to the environment

Result
------
A much easier way to configure global/default timeouts through the
spark config.

Change-Id: I4322df3e793ef2077fefcad4bf165b96f5f07bf4
Reviewed-on: http://review.couchbase.org/83495
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: David Nault <david.nault@couchbase.com>