SPARKC-77: add per-op timeout overloads.
Motivation
----------
In addition to easily configure the global timeout it also should
be possible to modify the timeout on a per-operation basis.

Modifications
-------------
This changeset builds on the previous global timeout foundation
and now exposes overloads to explicitly provide a duration (scala)
or simple long (java) in millis.

Result
------
Users can now provide per-operation timeout values as needed.

Change-Id: Ic8d7871694910db7637a7607d80f84183c7e15da
Reviewed-on: http://review.couchbase.org/83500
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: David Nault <david.nault@couchbase.com>