SPARKC-88: Allow N1QL queries to run on Spark node(s) co-located with query service
Motivation
----------
A N1QL query submitted to Spark will currently end up on a random Spark worker node.
To support Spark workers being co-located on Couchbase nodes, it would be better to
find the Couchbase nodes that are running query services, and try to run the query
on those Spark workers.

Modifications
----------
Tweaked QueryRDD to check the Couchbase config to find nodes that are running the
query service.  Then return a partition that contains the hostnames of those nodes.

Result
----------
Tested manually on a 3-node Spark cluster, with Spark nodes co-located on Couchbase
nodes.  N1QL queries were successfully sent to the Spark worker that was on the single
query node.

Change-Id: Idc16e9aaf086d1825a19a327b39989d790ad5ffc
Reviewed-on: http://review.couchbase.org/97854
Reviewed-by: David Nault <david.nault@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Michael Nitschinger <michael@nitschinger.at>