SPARKC-82: N1QLRelation should refuse to infer bucket name when there is more than 1 bucket
When N1QLRelation is created with a null bucket name, it uses the first bucket defined
in the config. The problem is that the bucket order is unspecified, since the buckets
are initially configured using the keys of a ConcurrentHashMap.

Contrast this with CouchbaseConnection.bucket, which throws an IllegalStateException
if the bucket name is not specified and there is not exactly one bucket.

For consistency, this change modifies N1QLRelation to behave similarly to
CouchbaseConnection.bucket.

Change-Id: I02f45cebbc21184ea36ece1e4277377dc6d9f8e2
Reviewed-on: http://review.couchbase.org/109178
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Graham Pople <grahampople@gmail.com>