SPARKC-166: Better handle aggregate datatypes.
This changeset fixes a couple issues around the result of aggregations
vs. their datatype from the original field. Consider the case where
an AVG is performed over a sequence of integers, the result is very likely
a fractional (i.e. float or double).

Also, the "complete aggregate pushdown" function is now enabled which allows
pushing down AVG into the connector.

Change-Id: Ic707e49fe52e9d2a206d8d8a608316b70514e598
Reviewed-on: https://review.couchbase.org/c/couchbase-spark-connector/+/181947
Reviewed-by: Graham Pople <graham.pople@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>