SPARKC-184: Spark Connector: Compile with Spark 3.5 SPARKC-183: Update dependencies
A generally painless upgrade from Spark 3.3 to 3.5,
and bumping all dependencies to the latest.

The connector duplicates some code from the Spark
codebase.  This needed to be synced with the latest
version of that code, and due to some privacy access
changes, needed to be moved into its original package.

Change-Id: I6b7b04e1be394f58f855945057dcb3bebb06b71d
Reviewed-on: https://review.couchbase.org/c/couchbase-spark-connector/+/202253
Tested-by: Graham Pople <graham.pople@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>