SPARKC-187: Spark Connector materializes all rows of query result
The connector currently buffers the full query result,
which can lead to OOM issues.

Replacing this with the reactive API.  Both the
PartitionReader Spark interface we need to extend
and reactor are pull-based, so we connect the two
together as such:

PartitionReader.next() will:
a) request more items from the reactive stream if needed.
b) block until there is a row available - or the stream
   ends or errors.

We try to maintain a queue of 30 items.  30 is chosen
arbitrarily, with the goal of striking a balance between
memory consumption and good throughput.

The algorithm will provide a backpressured stream of items.
If the Spark consumer is slow, items won't be requested
from the reactive stream.

Change-Id: Ib5052fba5f4d4946f923578c34192540e7efaa6e
Reviewed-on: https://review.couchbase.org/c/couchbase-spark-connector/+/202680
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Tested-by: Graham Pople <graham.pople@couchbase.com>
(cherry picked from commit 35cf6419b5a6e1f9337ca9d9620f59c810a2220d)
Reviewed-on: https://review.couchbase.org/c/couchbase-spark-connector/+/202313
Reviewed-by: Graham Pople <graham.pople@couchbase.com>