Tests: Fixing AlternateAddressIntegrationTest
This test was too-hastily de-testcontained when the
other tests were.

This particular test requires a cluster to have
alternate addresses configured - which clusters
running inside testcontainers happen to be.

Restoring the testcontainer for this test only.

(This is part of a chain of commits bringing 3.3
branch up to speed with the main 3.5 branch.)

Change-Id: Icdc273045db940a351d5b3e4f9da9385b08d2c9d
Reviewed-on: https://review.couchbase.org/c/couchbase-spark-connector/+/229107
Reviewed-by: Graham Pople <graham.pople@couchbase.com>
Tested-by: Graham Pople <graham.pople@couchbase.com>