Tests: Simplification #1
The existing TestInfra and SparkTest classes, used by
all tests, are intended to be a one-stop simple-to-use
base to help tests get bootstrapped without all the
C&P boilerplate they used to contain.

However, it's a flawed approach.   It couples together
too tightly SparkSession creation with test resource
creation, so any test that needs to do something outside
of the norm has to jump through hoops.  And it's overly
complicated.

So, first steps in moving to a simpler setup here.
The new classes are much more SOLID, and are designed
so tests setup only what they need with a few lines of
code.

SparkSimpleTest is provided, and this like the existing
TestInfra/SparkTest classes continues to offer a very
regimented setup that is suitable for many tests.  But
the difference is the underlying building blocks it uses
are simpler, clearer, and easier to C&P into a test
to get a more bespoke setup.

In other words - moving from framework to library.

The tests will be slowly ported to this new setup.

Change-Id: I2bf7b34984ff013e0eeeaa0acb50e75b9c8f0fdd
Reviewed-on: https://review.couchbase.org/c/couchbase-spark-connector/+/215253
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Graham Pople <graham.pople@couchbase.com>