SPARKC-212: First and last manual partitions are significantly larger
The partitioning logic was borrowed from the Spark JDBC
connector, which prioritises ensuring that most partitions
are the same size, at the expense of the first and last
partitions potentially being significantly larger.

As performance testing is showing this to have some
impact, changing the logic so that all partitions are now
approximately the same size, +/- 1 row.

Additionally adding some sanity checking for bugs.

Change-Id: I62a3815fb71264f58d9ee628c760df0d1f97ddf3
Reviewed-on: https://review.couchbase.org/c/couchbase-spark-connector/+/220164
Reviewed-by: Graham Pople <graham.pople@couchbase.com>
Tested-by: Graham Pople <graham.pople@couchbase.com>