Internal: remove C&Ped code
The prior maintainers copied some code from Spark,
presumably to obtain access to JSONOptionsInRead.

Using an alternate approach.

Some unused Columnar write code is using JSONOptionsInRead
directly.  Since Columnar writes are not presently
supporting, simply removing that chain of code for now.

(This is part of a chain of commits bringing 3.3
branch up to speed with the main 3.5 branch.)

Change-Id: I9bc309ef8cf7ce9661164053652f86ad9314608a
Reviewed-on: https://review.couchbase.org/c/couchbase-spark-connector/+/229119
Tested-by: Graham Pople <graham.pople@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>