Syncing 3.3 branch from master (3.5)
To aid development, aligning these branches as
much as possible.

Most changes have been cherrypicked already.  These
are the small differences that remain between the
branches.

Many are minor 2.13 changes that aren't needed
for 3.3, but will make it easier to maintain if the
branches are near-identical.

(This is part of a chain of commits bringing 3.3
branch up to speed with the main 3.5 branch.)

Change-Id: I88fe7019f859fc1b01b6e1d45ce8cd0182d4dd40
Reviewed-on: https://review.couchbase.org/c/couchbase-spark-connector/+/229118
Tested-by: Graham Pople <graham.pople@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>