SPARKC-230: Add failure reporting and handling for individual operations
Currently the failure of an individual KV write operation
inside a Spark DataFrame write, causes the entire write
to fail.

Adding two modes to help with this:

1. Add `df.write.format("couchbase.kv")
            .option(KeyValueOptions.ErrorHandler, "com.acme.YourErrorHandler")
            .save()`

This lets users provide a custom error handler, which
can allow the overall job to continue.  A sample handler
is also provided.
However this approach has significant limitations as it
will run on the Spark executor nodes and the user's
application will not have easy access to the information,
so support is also added for:

2. Add `df.write.format("couchbase.kv")
            `.option(KeyValueOptions.ErrorBucket, "error-bucket")`
            .save()`

This will write a document per failure, back to Couchbase.
The scope and collection can also be specified.

Change-Id: I60040a0fd8f558ed6252a2dbdd4ba5d2cb05e4aa
Reviewed-on: https://review.couchbase.org/c/couchbase-spark-connector/+/229187
Tested-by: Graham Pople <graham.pople@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>
(cherry picked from commit 321c81879ef3720804c9e4038f239d754146f507)
Reviewed-on: https://review.couchbase.org/c/couchbase-spark-connector/+/229579
Reviewed-by: Graham Pople <graham.pople@couchbase.com>